read the actual superblock from disk

+++ usr/xv6/kernel/fs/src/block.rs
@@ -34,17 +34,17 @@ impl Block {
             let mut buffer = bguard.lock();
 
             let mut bi = 0;
-            while bi < params::BPB && b + bi < super_block.size {
+            while bi < params::BPB && b + (bi as u32) < super_block.size {
                 let m = 1 << (bi % 8);
-                if buffer.data[bi as usize / 8] & m == 0 {
-                    buffer.data[bi as usize / 8] |= m; // mark block as used
+                if buffer.data[bi / 8] & m == 0 {
+                    buffer.data[bi / 8] |= m; // mark block as used
                     // TODO: log_write here
 
                     drop(buffer);
                     BCACHE.release(&mut bguard);
 
-                    Block::zero(device, (b + bi) as u32);
-                    return Some((b + bi) as u32);
+                    Block::zero(device, b + bi as u32);
+                    return Some(b + bi as u32);
                 }
                 bi += 1;
             }
